+++ packages/google/src/google-generative-ai-language-model.ts
@@ -1,5 +1,4 @@
 import {
-  InvalidArgumentError,
   LanguageModelV1,
   LanguageModelV1CallWarning,
   LanguageModelV1FinishReason,
@@ -14,9 +13,9 @@ import {
   combineHeaders,
   createEventSourceResponseHandler,
   createJsonResponseHandler,
+  parseProviderOptions,
   postJsonToApi,
   resolve,
-  safeValidateTypes,
 } from '@ai-sdk/provider-utils';
 import { z } from 'zod';
 import { convertJSONSchemaToOpenAPISchema } from './convert-json-schema-to-openapi-schema';
@@ -86,22 +85,13 @@ export class GoogleGenerativeAILanguageModel implements LanguageModelV1 {
 
     const warnings: LanguageModelV1CallWarning[] = [];
 
-    // parse and validate provider options:
-    const parsedProviderOptions =
-      providerMetadata != null
-        ? safeValidateTypes({
-            value: providerMetadata,
-            schema: providerOptionsSchema,
-          })
-        : { success: true as const, value: undefined };
-    if (!parsedProviderOptions.success) {
-      throw new InvalidArgumentError({
-        argument: 'providerOptions',
-        message: 'invalid provider options',
-        cause: parsedProviderOptions.error,
-      });
-    }
-    const googleOptions = parsedProviderOptions.value?.google;
+    const googleOptions = parseProviderOptions({
+      provider: 'google',
+      providerOptions: providerMetadata,
+      schema: z.object({
+        responseModalities: z.array(z.enum(['TEXT', 'IMAGE'])).nullish(),
+      }),
+    });
 
     const generationConfig = {
       // standardized settings:
@@ -630,14 +620,6 @@ const chunkSchema = z.object({
       promptTokenCount: z.number().nullish(),
       candidatesTokenCount: z.number().nullish(),
       totalTokenCount: z.number().nullish(),
-    })
-    .nullish(),
-});
-
-const providerOptionsSchema = z.object({
-  google: z
-    .object({
-      responseModalities: z.array(z.enum(['TEXT', 'IMAGE'])).nullish(),
     })
     .nullish(),
 });