+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,26 @@
 
 ### Patch Changes
 
+- Updated dependencies [5d0fc29]
+  - ai@4.1.66
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [16c444f]
+  - ai@4.1.65
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@4.1.64
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [0bd5bc6]
   - ai@4.1.63