+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/amazon-bedrock
 
+## 2.1.6
+
+### Patch Changes
+
+- Updated dependencies [d0c4659]
+  - @ai-sdk/provider-utils@2.1.15
+
 ## 2.1.5
 
 ### Patch Changes