+++ packages/deepseek/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/deepseek
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [d0c4659]
+  - @ai-sdk/provider-utils@2.1.15
+  - @ai-sdk/openai-compatible@0.1.17
+
 ## 0.1.16
 
 ### Patch Changes